+++ framework-crates/objc2-metal/tests/null_error_handling.rs
@@ -1,0 +1,25 @@
+//! Regression test for https://github.com/madsmtm/objc2/issues/653.
+#![cfg(all(feature = "MTLBinaryArchive", feature = "MTLDevice"))]
+use objc2_foundation::{ns_string, NSURL};
+use objc2_metal::{
+    MTLBinaryArchive, MTLBinaryArchiveDescriptor, MTLCreateSystemDefaultDevice, MTLDevice,
+};
+
+#[link(name = "CoreGraphics", kind = "framework")]
+extern "C" {}
+
+#[test]
+fn test() {
+    let Some(device) = MTLCreateSystemDefaultDevice() else {
+        // Ignore, this won't work in CI.
+        return;
+    };
+    let binary_archive = device
+        .newBinaryArchiveWithDescriptor_error(&MTLBinaryArchiveDescriptor::new())
+        .unwrap();
+
+    let metal_lib = unsafe { NSURL::URLWithString(ns_string!("file://missing.metallib")) }.unwrap();
+    let err = binary_archive.serializeToURL_error(&metal_lib).unwrap_err();
+    assert_eq!(err.domain().to_string(), "__objc2.missingError");
+    assert_eq!(err.code(), 0);
+}